organize utils; fix up unit tests
