+++ services/ui-src/src/components/index.tsx
@@ -1,8 +1,7 @@
-import { App } from "./wrappers/App";
-import { AppRoutes } from "./wrappers/AppRoutes";
+import { App } from "./app/App";
+import { AppRoutes } from "./app/AppRoutes";
 import { Header } from "./layout/Header";
 import { LoginCognito } from "./logins/LoginCognito";
 import { LoginIDM } from "./logins/LoginIDM";
-import { QueryProvider } from "./wrappers/QueryProvider";
 
-export { App, AppRoutes, Header, LoginCognito, LoginIDM, QueryProvider };
+export { App, AppRoutes, Header, LoginCognito, LoginIDM };
