+++ services/ui-src/src/custom.d.ts
@@ -1,0 +1,2 @@
+declare module "@cmsgov/design-system/dist/components/UsaBanner/UsaBanner";
+declare module "*.scss";
