+++ services/ui-src/src/index.tsx
@@ -4,10 +4,10 @@ import { ReactQueryDevtools } from "react-query/devtools";
 import { Amplify } from "aws-amplify";
 import config from "config";
 // utils
-import * as serviceWorker from "utils/serviceWorker";
-import { UserProvider, ApiProvider } from "hooks/authHooks";
+import { ApiProvider, QueryProvider } from "utils/api";
+import { UserProvider } from "utils/auth";
 // components
-import { App, QueryProvider } from "components";
+import { App } from "components";
 // styles
 import { ChakraProvider } from "@chakra-ui/react";
 import { theme } from "styles/theme";
@@ -36,10 +36,3 @@ ReactDOM.render(
   </Router>,
   document.getElementById("root")
 );
-
-/*
- * If you want your app to work offline and load faster, you can change
- * unregister() to register() below. Note this comes with some pitfalls.
- * Learn more about service workers: https://bit.ly/CRA-PWA
- */
-serviceWorker.unregister();
