--- services/ui-src/src/libs/contextLib.ts
@@ -1,12 +1,0 @@
-import React, { useContext, createContext } from "react";
-
-export const AppContext = createContext<IAppContextInterface>({});
-
-export function useAppContext() {
-  return useContext(AppContext);
-}
-
-interface IAppContextInterface {
-  isAuthenticated?: boolean;
-  userHasAuthenticated?: React.Dispatch<React.SetStateAction<boolean>>;
-}
