--- services/ui-src/src/libs/errorLib.ts
@@ -1,10 +1,0 @@
-export function onError(error: Error | any): void {
-  let message = error.toString();
-
-  // Auth errors
-  if (!(error instanceof Error) && error.message) {
-    message = error.message;
-  }
-
-  alert(message);
-}
