--- services/ui-src/src/libs/index.tsx
@@ -1,8 +1,0 @@
-export * as Api from "./api";
-export * from "./awsLib";
-export * from "./contacts";
-export * from "./contextLib";
-export * from "./errorLib";
-export * from "./helpers";
-export * from "./hooksLib";
-export * from "./territoryLib";
