--- services/ui-src/src/libs/spaLib.ts
@@ -1,224 +1,0 @@
-export const SPA = [
-  {
-    state: "California",
-    name: "California Health Home Program (HHP)",
-    id: "CA 19-0037",
-    postal: "CA",
-  },
-  {
-    state: "California",
-    name: "California Behavioral Health Home",
-    id: "CA 20-002",
-    postal: "CA",
-  },
-  {
-    state: "Connecticut",
-    name: "Serious and Persistent Mental Illness",
-    id: "CT 15-014",
-    postal: "CT",
-  },
-  {
-    state: "Delaware",
-    name: "Delaware Assertive Community Integration Support Team (ACIST) Health Home",
-    id: "DE 18-006",
-    postal: "DE",
-  },
-  {
-    state: "District of Columbia",
-    name: "Chronic Care Management for Individuals with Serious and Persistent Mental Health Conditions",
-    id: "DC 18-0006",
-    postal: "DC",
-  },
-  {
-    state: "District of Columbia",
-    name: "Chronic Conditions",
-    id: "DC 18-0007",
-    postal: "DC",
-  },
-  {
-    state: "Iowa",
-    name: "Iowa Health Home Services",
-    id: "IA 20-0010",
-    postal: "IA",
-  },
-  {
-    state: "Iowa",
-    name: "Iowa Severe and Persistent Mental Illness Health Home ",
-    id: "IA 20-0011",
-    postal: "IA",
-  },
-  {
-    state: "Kansas",
-    name: "Health Home Serving Serious Mental Illness (SMI)",
-    id: "KS 20-0004",
-    postal: "KS",
-  },
-  {
-    state: "Kansas",
-    name: "Health Home Chronic Conditions - Asthma",
-    id: "KS 20-0005",
-    postal: "KS",
-  },
-  {
-    state: "Maine",
-    name: "Maine Stage A Health Home Targeting Individuals with Chronic Conditions",
-    id: "ME 13-012",
-    postal: "ME",
-  },
-  {
-    state: "Maine",
-    name: "Maine Behavioral Health Home",
-    id: "ME 18-0002",
-    postal: "ME",
-  },
-  {
-    state: "Maine",
-    name: "Maine Health Home for Beneficiaries Receiving Medication Assisted Treatment for Opioid Addiction",
-    id: "ME 18-0032",
-    postal: "ME",
-  },
-  {
-    state: "Maryland",
-    name: "Maryland Health Home Services",
-    id: "MD 20-0006",
-    postal: "MD",
-  },
-  {
-    state: "Michigan",
-    name: "Chronic Care Model for Individuals with Serious and Persistent Mental Health Conditions",
-    id: "MI 16-1500",
-    postal: "MI",
-  },
-  {
-    state: "Michigan",
-    name: "Michigan Care Team",
-    id: "MI 15-2000",
-    postal: "MI",
-  },
-  {
-    state: "Michigan",
-    name: "Opioid Health Home",
-    id: "MI 18-1500",
-    postal: "MI",
-  },
-  {
-    state: "Minnesota",
-    name: "Minnesota Behavioral Health Home",
-    id: "MN 19-0015",
-    postal: "MN",
-  },
-  {
-    state: "Missouri",
-    name: "Community Mental Health Center Health Home",
-    id: "MO 19-0020",
-    postal: "MO",
-  },
-  {
-    state: "Missouri",
-    name: "Primary Care Clinic Health Home",
-    id: "MO 19-0003",
-    postal: "MO",
-  },
-  {
-    state: "New Jersey",
-    name: "New Jersey Behavioral Health Home (Adults)",
-    id: "NJ 16-0001",
-    postal: "NJ",
-  },
-  {
-    state: "New Jersey",
-    name: "New Jersey Behavioral Health Home (Children)",
-    id: "NJ 16-0002",
-    postal: "NJ",
-  },
-  {
-    state: "New Mexico",
-    name: "CareLink New Mexico",
-    id: "NM 18-0002",
-    postal: "NM",
-  },
-  {
-    state: "New York",
-    name: "Health Home for High-Cost, High-Needs Enrollees ",
-    id: "NY 20-0054",
-    postal: "NY",
-  },
-  {
-    state: "New York",
-    name: "New York I/DD Health Home Services",
-    id: "NY 17-0025",
-    postal: "NY",
-  },
-  {
-    state: "Oklahoma",
-    name: "Oklahoma Health Home (Adults)",
-    id: "OK 14-0012",
-    postal: "OK",
-  },
-  {
-    state: "Oklahoma",
-    name: "Oklahoma Health Home (Children)",
-    id: "OK 14-0011",
-    postal: "OK",
-  },
-  {
-    state: "Rhode Island",
-    name: "Comprehensive Evaluation, Diagnosis, Assessment, Referral, & Reevaluation (CEDARR) Family Centers Health Home",
-    id: "RI 18-009",
-    postal: "RI",
-  },
-  {
-    state: "Rhode Island",
-    name: "Community Mental Health Organizations Health Home",
-    id: "RI 18-006",
-    postal: "RI",
-  },
-  {
-    state: "Rhode Island",
-    name: "Rhode Island Opioid Treatment Program Health Home Services",
-    id: "RI 16-006",
-    postal: "RI",
-  },
-  {
-    state: "South Dakota",
-    name: "South Dakota Health Home",
-    id: "SD 20-0008",
-    postal: "SD",
-  },
-  {
-    state: "Tennessee",
-    name: "Tennessee Health Link Health Home Program ",
-    id: "TN 16-004",
-    postal: "TN",
-  },
-  {
-    state: "Vermont",
-    name: "Vermont Health Home for Beneficiaries Receiving Medication Assisted Treatment for Opioid Addiction ",
-    id: "VT 14-007",
-    postal: "VT",
-  },
-  {
-    state: "Washington",
-    name: "Washington Health Home Services",
-    id: "WA 20-0031",
-    postal: "WA",
-  },
-  {
-    state: "West Virginia",
-    name: "West Virginia Health Home for Individuals with Bipolar Disorder at Risk for Hepatitis Type B and C",
-    id: "WV 16-0007",
-    postal: "WV",
-  },
-  {
-    state: "West Virginia",
-    name: "West Virginia Health Home for Individuals with Chronic Conditions",
-    id: "WV 16-0008",
-    postal: "WV",
-  },
-  {
-    state: "Wisconsin",
-    name: "Individuals with HIV/AIDS",
-    id: "WI 12-008",
-    postal: "WI",
-  },
-];
