+++ services/ui-src/src/utils/api/index.ts
@@ -1,3 +1,5 @@
 export * from "./useGetMeasures";
 export * from "./useCreateMeasure";
 export * from "./useDeleteMeasure";
+export * from "./ApiProvider";
+export * from "./QueryProvider";
