+++ services/ui-src/src/utils/auth/index.ts
@@ -1,3 +1,2 @@
 export * from "./useUser";
 export * from "./userProvider";
-export * from "./ApiProvider";
