--- services/ui-src/src/utils/testing.tsx
@@ -1,16 +1,0 @@
-import React from "react";
-import { BrowserRouter as Router } from "react-router-dom";
-
-jest.mock("react-router-dom", () => ({
-  ...jest.requireActual("react-router-dom"),
-  useParams: () => ({
-    year: "2021",
-    state: "OH",
-    coreSetId: "ACS",
-    measureId: "AIF-HH",
-  }),
-}));
-
-export const RouterWrappedComp: React.FC = ({ children }) => (
-  <Router>{children}</Router>
-);
