+++ services/ui-src/src/utils/testing/mockAsset.ts
