+++ services/ui-src/src/views/Home/Home.test.tsx
@@ -1,8 +1,9 @@
 import { render, screen } from "@testing-library/react";
 import { QueryClient, QueryClientProvider } from "react-query";
+import { axe } from "jest-axe";
 // utils
-import { RouterWrappedComp } from "utils/testing";
-import { useApiMock } from "utils/testUtils/useApiMock";
+import { RouterWrappedComponent } from "utils/testing/setupJest";
+import { useApiMock } from "utils/testing/mockApi";
 // views
 import Home from "./Home";
 
@@ -13,14 +14,28 @@ describe("Test Home view", () => {
     useApiMock({});
     render(
       <QueryClientProvider client={queryClient}>
-        <RouterWrappedComp>
+        <RouterWrappedComponent>
           <Home />
-        </RouterWrappedComp>
+        </RouterWrappedComponent>
       </QueryClientProvider>
     );
   });
 
-  test("Check that the Home renders", () => {
+  test("Check that Home view renders", () => {
     expect(screen.getByTestId("home-view")).toBeVisible();
+  });
+});
+
+describe("Test Home view accessibility", () => {
+  it("Should not have basic accessibility issues", async () => {
+    const { container } = render(
+      <QueryClientProvider client={queryClient}>
+        <RouterWrappedComponent>
+          <Home />
+        </RouterWrappedComponent>
+      </QueryClientProvider>
+    );
+    const results = await axe(container);
+    expect(results).toHaveNoViolations();
   });
 });
